LayoutTests/fast/text/text-orientation-parse.html

<!DOCTYPE html>
<html>
	<head>
	<script src="../../resources/js-test-pre.js"></script>
	</head>
	<body>
		<div style="-webkit-writing-mode: vertical-rl; text-orientation: sideways;">
			<div id="test1"></div>
			<div id="test2" style="-webkit-writing-mode: vertical-rl; text-orientation: sideways;"></div>
			<div id="test3" style="-webkit-writing-mode: vertical-rl; text-orientation: sideways;"></div>
			<div id="test4" style="-webkit-writing-mode: vertical-rl; text-orientation: sideways;"></div>
			<div id="test5" style="-webkit-writing-mode: vertical-rl; text-orientation: upright;">
				<div id="test6"></div>
				<div id="test7" style="text-orientation: upright;">
					<div id="test8">
						<div id="test9"></div>
					</div>
					<div id="test10" style="text-orientation: mixed;"></div>
				</div>
			</div>
			<div id="test11" style="-webkit-text-orientation: upright; text-orientation: sideways;"></div> 
		</div>


		<div id="test12" style="-webkit-writing-mode: vertical-rl;">
			<div id="test13"></div>
		</div>


		<div style="-webkit-writing-mode: vertical-rl; -webkit-text-orientation: sideways;">
			<div id="test14"></div>
			<div id="test15" style="-webkit-writing-mode: vertical-rl; text-orientation: sideways;"></div>
			<div id="test16" style="-webkit-writing-mode: vertical-rl; text-orientation: upright;">
				<div id="test17"></div>
				<div id="test18" style="text-orientation: upright;">
					<div id="test19">
						<div id="test20"></div>
					</div>
					<div id="test21" style="text-orientation: mixed;"></div>
				</div>
			<div id="test22" style="-webkit-text-orientation: upright; text-orientation: sideways;"></div> 
			</div>
		</div>


		<div style="-webkit-writing-mode: vertical-rl; -webkit-text-orientation: sideways; text-orientation: upright">
			<div id="test23">
				<div id="test24"></div>
			</div>
		</div>

		<div id="test25" style="-webkit-writing-mode: vertical-rl;  text-orientation: upright; -webkit-text-orientation: sideways;">
			<div id="test26">
				<div id="test27"></div>
			</div>
		</div>



		<script>
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test7')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test8')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test9')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test10')).webkitTextOrientation", "mixed");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test11')).webkitTextOrientation", "sideways");


			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test12')).webkitTextOrientation", "mixed");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test13')).webkitTextOrientation", "mixed");


			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test14')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test15')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test16')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test17')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test18')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test19')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test20')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test21')).webkitTextOrientation", "mixed");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test22')).webkitTextOrientation", "sideways");

			
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test23')).webkitTextOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test24')).webkitTextOrientation", "upright");

			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test25')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test26')).webkitTextOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test27')).webkitTextOrientation", "sideways");


			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test7')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test8')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test9')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test10')).textOrientation", "mixed");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test11')).textOrientation", "sideways");


			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test12')).textOrientation", "mixed");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test13')).textOrientation", "mixed");


			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test14')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test15')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test16')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test17')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test18')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test19')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test20')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test21')).textOrientation", "mixed");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test22')).textOrientation", "sideways");

			
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test23')).textOrientation", "upright");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test24')).textOrientation", "upright");

			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test25')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test26')).textOrientation", "sideways");
			shouldBeEqualToString("window.getComputedStyle(document.getElementById('test27')).textOrientation", "sideways");


		</script>

		<script src="../../resources/js-test-post.js"></script>
	</body>
</html>
