Safari-611.1.10/LayoutTests/fast/text/text-orientation-parse-expected.txt

PASS window.getComputedStyle(document.getElementById('test1')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test2')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test3')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test4')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test5')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test6')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test7')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test8')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test9')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test10')).webkitTextOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test11')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test12')).webkitTextOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test13')).webkitTextOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test14')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test15')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test16')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test17')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test18')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test19')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test20')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test21')).webkitTextOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test22')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test23')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test24')).webkitTextOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test25')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test26')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test27')).webkitTextOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test1')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test2')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test3')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test4')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test5')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test6')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test7')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test8')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test9')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test10')).textOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test11')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test12')).textOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test13')).textOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test14')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test15')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test16')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test17')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test18')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test19')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test20')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test21')).textOrientation is "mixed"
PASS window.getComputedStyle(document.getElementById('test22')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test23')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test24')).textOrientation is "upright"
PASS window.getComputedStyle(document.getElementById('test25')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test26')).textOrientation is "sideways"
PASS window.getComputedStyle(document.getElementById('test27')).textOrientation is "sideways"
PASS successfullyParsed is true

TEST COMPLETE
